LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-scroll-anchoring/editable-scroll-anchor-expected.txt



FAIL Scroll anchoring is suppressed during editing in a textarea assert_equals: expected 30 but got 94
PASS Scroll anchoring is suppressed during editing in a contenteditable element
